lb_bench.py: Trace newidle_balance instead of load_balance
The original idea was to trace the load_balance calls from
newidle_balance. For that, we used the approach to check the stack
trace. Unfortunately, the stack traces are inconsistent across diferent
kernel versions.

We now trace the newidle_balance calls, and because of that we don't
need the strack traces anymore.